backend/requirements.txt
Flask-Cors
Flask
requests
python-dotenv
langchain>=0.1.0
langchain-community>=0.1.0
langchain-experimental>=0.1.0
pydantic>=2.0.0